♻️ fix copy attribute & move use-item-group to lib/hooks
