+++ apps/web/ui/entity/overview/entry.tsx
@@ -4,12 +4,13 @@ import * as React from "react";
 import { Status } from "~/ui/status";
 import { CopyableValue } from "~/ui/copyable-value";
 
-import { useItemGrid } from "~/ui/search/use-item-grid";
+import { useItemGrid } from "~/lib/hooks/use-item-grid";
 import { cn } from "~/ui/shadcn/utils";
 import { copyValueToClipboard, truncateHash } from "~/lib/shared-utils";
 import { toast } from "~/ui/shadcn/components/ui/use-toast";
 
 import type { Value } from "@modularcloud/headless";
+import { useHotkeyListener } from "~/lib/hooks/use-hotkey-listener";
 
 interface Props {
   entries: Array<[key: string, value: Value]>;
@@ -35,6 +36,29 @@ export function OverviewEntryList({ entries }: Props) {
     onSelectOption: ({ value }) => {
       /** TODO: should navigate to link if value type is `link` */
     },
+  });
+
+  useHotkeyListener({
+    keys: ["c"],
+    listener: async () => {
+      if (!selectedOption) return;
+      const { type, payload } = selectedOption.value;
+      if (payload === null || payload === undefined) return;
+
+      if (type === "standard" || type === "longval") {
+        const value =
+          type === "standard" ? payload.toString() : payload.value.toString();
+        const copied = await copyValueToClipboard(value);
+
+        if (copied) {
+          toast({
+            title: "Copied",
+            description: `"${truncateHash(value)}" copied to clipboard`,
+          });
+        }
+      }
+    },
+    modifier: "META",
   });
 
   return (
@@ -129,20 +153,20 @@ export function OverviewEntry({
         },
       )}
       {...optionProps}
-      onCopy={async () => {
-        if (type === "standard" || type === "longval") {
-          const value =
-            type === "standard" ? payload.toString() : payload.value.toString();
-          const copied = await copyValueToClipboard(value);
+      // onCopy={async () => {
+      //   if (type === "standard" || type === "longval") {
+      //     const value =
+      //       type === "standard" ? payload.toString() : payload.value.toString();
+      //     const copied = await copyValueToClipboard(value);
 
-          if (copied) {
-            toast({
-              title: "Copied",
-              description: `"${truncateHash(value)}" copied to clipboard`,
-            });
-          }
-        }
-      }}
+      //     if (copied) {
+      //       toast({
+      //         title: "Copied",
+      //         description: `"${truncateHash(value)}" copied to clipboard`,
+      //       });
+      //     }
+      //   }
+      // }}
     >
       <dt className="col-span-2 font-medium">{label}</dt>
